fix: 🐛 show spinner even on no map view
BREAKING CHANGE: 🧨 None
