+++ src/pages/Search.tsx
@@ -4,8 +4,6 @@ import MainLayout from 'src/layouts/main';
 import { Box, styled } from '@mui/material';
 import { SearchForm } from 'src/containers/search/SearchForm';
 import { HEADER } from 'src/config/componentSizes';
-import { useAccommodationsAndOffers } from '../hooks/useAccommodationsAndOffers.tsx';
-import { InvalidLocationError } from '../hooks/useAccommodationsAndOffers.tsx/helpers';
 
 const SearchBox = styled(Box)(({ theme }) => ({
   position: 'absolute',
@@ -19,28 +17,13 @@ const SearchBox = styled(Box)(({ theme }) => ({
   }
 }));
 
-const NoMapBox = styled(Box)(({ theme }) => ({
-  width: '100%',
-  height: '100%',
-  backgroundColor: theme.palette.background.default
-}));
-
 //TODO: remove the footer
 export const Search = () => {
-  const { error } = useAccommodationsAndOffers();
-  const invalidLocation = error instanceof InvalidLocationError;
   return (
     <MainLayout childrenBelowHeader={false} footer={false}>
       <Box style={{ position: 'relative' }}>
-        {invalidLocation ? (
-          <NoMapBox />
-        ) : (
-          <>
-            <Results />
-            <MapBox />
-          </>
-        )}
-
+        <Results />
+        <MapBox />
         <SearchBox>
           <SearchForm />
         </SearchBox>
